[refac] ajuste no listUsers
